server/requirements.txt
Flask==2.3.2
PyJWT==2.9.0
